Adding in code parsing with syntax shown in example.cpp
this is a combination of the following commits:
Adding initial gui textbox at bottom of screen.

adding simple textbox gui in curses with python.

added in the ability to display messages.

example cpp file added in and gui updated to print code in top box.

Now able to parse code.

Removed gui.cpp